Parameter Conversions

This page is under construction, thank you for your patience.

We provide helper functions to convert between various parameters commonly used to describe the observing conditions.

Helper Functions

Kolmogorov.Cn_squared_to_fried — Function

Cn_squared_to_fried(Cn_squared::AbstractFloat,
                    wavelength::AbstractFloat=500e-9)

Calculates the Fried parameter ($r_0$) from a provided $C_n^2$ value and (optionally) wavelength in meters. If no wavelength is provided, the value assumed is 500 nm. The returned Fried parameter has units of meters.

Kolmogorov.fried_to_Cn_squared — Function

fried_to_Cn_squared(r0::AbstractFloat,
                    wavelength::AbstractFloat=500e-9)

Calculates the integrated $C_n^2$ value that yields the provided Fried parameter ($r_0$). If wavelength is not provided, the value assumed is 500 nm. The returned $C_n^2$ has units of $\mathrm{m}^{-2/3}$.

Kolmogorov.fried_to_seeing — Function

fried_to_seeing(r0::AbstractFloat,
                wavelength::AbstractFloat=500e-9)

Calculates the resulting seeing from a given Fried parameter ($r_0$) and (optionally) wavelength, both in meters. If no wavelength is provided, the value assumed is 500 nm. The returned seeing is in arcseconds.

Kolmogorov.seeing_to_fried — Function

seeing_to_fried(seeing::AbstractFloat,
                wavelength::AbstractFloat=500e-9)

Calculates the Fried parameter ($r_0$) that yields the provided seeing, seeing must be in arcseconds. If no wavelength is provided, the value assumed is 500 nm. The returned Fried parameter is in units of meters.