Standard Atmospheric Statistics

Here we go through the standard set of atmosphere statistics which are commonly used in adaptive optics simulations.

Hermitian Noise

The phase screens generated use the approach generating Gaussian Hermitian White Noise, which is then filtered to have a power spectrum matching a particular type. At this point, the inverse FFT is applied to get the phase screen itself. Gaussian Hermitian white noise is generated from the FFT of real white noise, as long as the filter applied to this data is strictly real then the hermitian nature will be preserved and the inverse FFT will still produce strictly real values.

Kolmogorov.generate_gaussian_hermitian_noise — Function

generate_gaussian_hermitian_noise(N::int)

Generates Gaussian hermitian white noise, specifically noise that when you take the inverse FFT the output is strictly real (aside from the imprecision of FFTs). Note that this means it is not a hermitian matrix, because the FFT indexing is different. Returned matrix is NxN and populated with complex numbers.

Kolmogorov

This page is under construction, thank you for your patience.

Kolmogorov.kolmogorov_psd — Function

kolmogorov_psd(frequencies::AbstractVector{<:AbstractFloat},
               Cn_squared::AbstractFloat)

Generates the 2D Kolmogorov power spectral density (PSD) with a given $C_n^2$ value which determines the overall strength. The PSD is evaluated on a grid defined by frequencies, which are assumed to be the angular spatial frequencies along one axis of the grid. Specifically, they should be in units of radians/meter. Cn_squared determines the overall normalization of the output PSD. The Kolmogorov PSD is defined as:

$\Phi_n\left(\vec{k}\right) = 0.033 \: C_n^2 \: \vec{k}^{-11/3}$

where $\vec{k} = 2\pi\left(f_x \: \hat{i} + f_y \: \hat{j}\right)$ is the angular spatial frequency in radians/meter

Missing docstring.

Missing docstring for Kolmogorob.kolmogorov_filter!. Check Documenter's build log for details.

von Karman

Missing docstring.

Missing docstring for Kolmogorov.vonkarman_psd. Check Documenter's build log for details.

Modified von Karman

Kolmogorov.modified_vonkarman_psd — Function

modified_vonkarman_psd(frequencies::AbstractVector{<:AbstractFloat},
                       Cn_squared::AbstractFloat,
                       L0::AbstractFloat,
                       l0::AbstractFloat)

Generates the 2D modified von Karman power spectral density (PSD) with a given $C_n^2$ value which determines the overall strength, and $L_0$ which determines the distance at which power rolls off at large distances, and $l_0$, which determines the power roll off at small distances. The PSD is evaluated on a grid defined by frequencies, which are assumed to be the angular spatial frequencies along one axis of the grid. Specifically, they should be in units of radians/meter. Cn_squared determines the overall normalization of the output PSD. L0 is the outer scale, which is, roughly speaking, the distance after which power no longer increases. The von Karman PSD is defined as:

$\Phi_n\left(\vec{k}\right) = \frac{0.033 \: C_n^2 \: e^{\left(k^2/k_m^2\right)}}{\left(\vec{k} \: + \: k_0\right)^{-11/3}} \:$

where $\vec{k} = 2\pi\left(f_x \: \hat{i} + f_y \: \hat{j}\right)$ is the angular spatial frequency in radians/meter, $k_0 = 2\pi/L_0$, and $k_m = 5.92/l_0$